Introduce synthetics and cloud domains (#102)
This change is a refectoring of a package structure for synthetics and
cloud exports code.

As the codebase grows larger, there are drawbacks of keeping packages split
to layers for all domains (models, resources, api_payloads, etc.).
For example, I found that there is a need to prefix model types with
the name of their domain for namespacing purposes - to avoid name conflicts.
Extracting models to domain packages (synthetics, cloud) fixes that problem and changes the usage
of exported types, e.g. models.SyntheticsAgent changes to synthetics.Agent.

The same applies to internal packages. It is easier to work with multiple
domain-specific packages instead of layer-based mega packages
(resources, api_payloads).

I propose to keep this domain-base style for all new APIs.

Issue: KNTK-399
Based-on: #99 (to avoid conflicts)